v1.3 Enhancements and Fixes
* Fixed an issue with the Page Bottom widget area when using a language
translation. Adjusted the sidebar-pagebottom.php to call the sidebar by
ID rather than name since language translations can change the name.
* Enhanced the CSS to better handle the WordPress Admin Bar when it's
in use. Previously, the main menu overlapped it a bit which was ugly
and made it harder to navigate.
* Updated the language translation tag for the social menu. It
previously had a different tag ('xtremelysocial') than the rest of the
theme ('flat-bootstrap').
* Changed a function name that referred to our spot theme instead of
this link theme. It didn't cause any problems, but was confusing to
anyone looking at the code.
* Removed reference to Blacktie.co from the site credits that
automatically display at the bottom of the theme. They are still
credited in this readme.txt file of course.