v1.4 enhancements and fixes
* Changed sidebar-pagebottom.php to always fire off the
xsbf_pagebottom() filter so that plugins can override what displays
there even if a user hasn't added any widgets yet.
* Load the parent theme's (Flat Bootstrap) style.css from functions.php
to improve performance instead of using @include in the theme's
style.css.
* Removed CSS for "after-footer" (footer nav menu and site credits) so
it matches the parent theme. This is needed for when Flat Bootstrap
v2.0 is released.
* Updated the comments in functions.php to reflect the new option for
custom header location (not used by this theme).
* We now completely override the parent theme's
xsbf_custom_header_setup() because when v2.0 of Flat Bootstrap is
released, it will have its own custom headers.
* If Jetpack Social Menu is activated, this theme will use that instead
of our own. We have also added the phrase "Jetpack" to their social
menu so its clear which one is theirs and which one is ours.
* Added footer.php to override the parent theme. This is needed for
when Flat Bootstrap v2.0 is released.
* Added a table of contents to style.css and better organized the
sections.